inserted RA/PR into database

+++ back-api-db/src/db/schema/create.sql
@@ -11,8 +11,8 @@ DROP TABLE IF EXISTS budget CASCADE;
 CREATE TABLE users (
   id SERIAL PRIMARY KEY NOT NULL,
   name VARCHAR(255) NOT NULL,
-  riskScore INTEGER,
-  portfolioReturn INTEGER
+  riskScore REAL,
+  portfolioReturn REAL
 );
 
 CREATE TABLE expenses (
